feat: auto switch indexer or common sync
